+++ tests/basic.rs
@@ -51,19 +51,31 @@ pub trait Kita {
 }
 
 const _: () = {
-    trait _Kita0<T0: ?Sized> {
+    pub trait _Kita0<_0: ?Sized> {
         const NAME: &'static str;
+
+        fn name() -> &'static str {
+            "Default blanket"
+        }
     }
 
-    impl<T0: Dispatch<Group = GroupA>> _Kita0<GroupA> for T0 {
+    impl<_0: Dispatch<Group = GroupA>> _Kita0<GroupA> for _0 {
         const NAME: &'static str = "Blanket A";
     }
-    impl<T0: Dispatch<Group = GroupB>> _Kita0<GroupB> for T0 {
+    impl<_0: Dispatch<Group = GroupB>> _Kita0<GroupB> for _0 {
         const NAME: &'static str = "Blanket B";
+
+        fn name() -> &'static str {
+            "Blanket B"
+        }
     }
 
-    impl<T0> Kita for T0 where T0: Dispatch, Self: _Kita0<<T0 as Dispatch>::Group> {
-        const NAME: &'static str = <Self as _Kita0<<T0 as Dispatch>::Group>>::NAME;
+    impl<_0> Kita for _0 where _0: Dispatch, Self: _Kita0<<_0 as Dispatch>::Group> {
+        const NAME: &'static str = <Self as _Kita0<<_0 as Dispatch>::Group>>::NAME;
+
+        fn name() -> &'static str {
+            <Self as _Kita0<<_0 as Dispatch>::Group>>::name()
+        }
     }
 };
 */