+++ tests/blanket_for_derivative_type.rs
@@ -39,21 +39,22 @@ pub trait Kita {
 }
 
 const _: () = {
-    trait _Kita0<T0: ?Sized> {
+    pub trait _Kita0<_0: ?Sized> {
         const NAME: &'static str;
     }
 
-    impl<T0: Dispatch<Group = GroupA>> _Kita0<GroupA> for Option<T0> {
+    impl<_1: Dispatch<Group = GroupA>> _Kita0<GroupA> for Option<_1> {
         const NAME: &'static str = "Blanket A";
     }
-    impl<T0: Dispatch<Group = GroupB>> _Kita0<GroupB> for Option<T0> {
+    impl<_1: Dispatch<Group = GroupB>> _Kita0<GroupB> for Option<_1> {
         const NAME: &'static str = "Blanket B";
     }
 
-    impl<T0> Kita for Option<T0> where T0: Dispatch, Self: _Kita0<<T0 as Dispatch>::Group> {
-        const NAME: &'static str = <Self as _Kita0<<T0 as Dispatch>::Group>>::NAME;
+    impl<_1> Kita for Option<_1> where _1: Dispatch, Self: _Kita0<<_1 as Dispatch>::Group> {
+        const NAME: &'static str = <Self as _Kita0<<_1 as Dispatch>::Group>>::NAME;
     }
 };
+
 */
 
 #[test]