+++ tests/disjoint_inherent_impl.rs
@@ -20,40 +20,40 @@ impl Dispatch for u32 {
     type Group = GroupB;
 }
 
-pub struct Wrapper<T>(T);
+pub struct Wrapper<'a, T, const N: usize>(&'a T);
 
 disjoint_impls! {
-    impl<T> Wrapper<T> where T: Dispatch<Group = GroupA> {
+    impl<T> Wrapper<'_, T, 12> where T: Dispatch<Group = GroupA> + Dispatch<Group = GroupA> {
         pub const NAME: &'static str = "Blanket A";
     }
-    impl<T> Wrapper<T> where T: Dispatch<Group = GroupB> {
+    impl<T> Wrapper<'_, T, 12> where T: Dispatch<Group = GroupB> {
         pub const NAME: &'static str = "Blanket B";
     }
 }
 
 /*
 const _: () = {
-    trait _Wrapper0<T0: ?Sized> {
+    pub trait _Wrapper0<_0: ?Sized> {
         const NAME: &'static str;
     }
 
-    impl<T0> _Wrapper0<GroupA> for Wrapper<T0> where T0: Dispatch<Group = GroupA> {
+    impl<_2> _Wrapper0<GroupA> for Wrapper<'_, _2, 12> where _2: Dispatch<Group = GroupA> + Dispatch<Group = GroupA> {
         const NAME: &'static str = "Blanket A";
     }
-    impl<T0> _Wrapper0<GroupB> for Wrapper<T0> where T0: Dispatch<Group = GroupB> {
+    impl<_2> _Wrapper0<GroupB> for Wrapper<'_, _2, 12> where _2: Dispatch<Group = GroupB> {
         const NAME: &'static str = "Blanket B";
     }
 
-    impl<T0> Wrapper<T0> where T0: Dispatch, Self: _Wrapper0<<T0 as Dispatch>::Group> {
-        const NAME: &'static str = <Self as _Wrapper0<<T0 as Dispatch>::Group>>::NAME;
+    impl<_2> Wrapper<'_, _2, 12> where _2: Dispatch, Self: _Wrapper0<<_2 as Dispatch>::Group> {
+        pub const NAME: &'static str = <Self as _Wrapper0<<_2 as Dispatch>::Group>>::NAME;
     }
 };
 */
 
 #[test]
 fn disjoint_inherent_impl() {
-    assert_eq!("Blanket A", <Wrapper<String>>::NAME);
-    assert_eq!("Blanket A", <Wrapper<Vec::<u32>>>::NAME);
-    assert_eq!("Blanket B", <Wrapper<u32>>::NAME);
-    assert_eq!("Blanket B", <Wrapper<i32>>::NAME);
+    assert_eq!("Blanket A", <Wrapper<String, 12>>::NAME);
+    assert_eq!("Blanket A", <Wrapper<Vec::<u32>, 12>>::NAME);
+    assert_eq!("Blanket B", <Wrapper<u32, 12>>::NAME);
+    assert_eq!("Blanket B", <Wrapper<i32, 12>>::NAME);
 }