+++ tests/multiple_blanket_impls.rs
@@ -25,6 +25,13 @@ disjoint_impls! {
         const NAME: &'static str;
     }
 
+    impl<T: Dispatch<Group = GroupA>> Kita for T {
+        const NAME: &'static str = "Blanket A";
+    }
+    impl<U: Dispatch<Group = GroupB>> Kita for U {
+        const NAME: &'static str = "Blanket B";
+    }
+
     impl<T: Dispatch<Group = GroupA>, U: Dispatch<Group = GroupA>> Kita for (T, U) {
         const NAME: &'static str = "Blanket AA";
     }
@@ -38,13 +45,6 @@ disjoint_impls! {
     impl<T: Dispatch<Group = GroupB>, U> Kita for (T, U) {
         const NAME: &'static str = "Blanket B*";
     }
-
-    impl<T: Dispatch<Group = GroupA>> Kita for T {
-        const NAME: &'static str = "Blanket A";
-    }
-    impl<U: Dispatch<Group = GroupB>> Kita for U {
-        const NAME: &'static str = "Blanket B";
-    }
 }
 
 /*
@@ -53,34 +53,34 @@ pub trait Kita {
 }
 
 const _: () = {
-    pub trait _Kita0<T0: ?Sized, T1: ?Sized> {
+    pub trait _Kita0<_0: ?Sized> {
         const NAME: &'static str;
     }
-    pub trait _Kita1<T0: ?Sized> {
+    pub trait _Kita1<_0: ?Sized, _1: ?Sized> {
         const NAME: &'static str;
     }
 
-    impl<T0: Dispatch<Group = GroupA>, T1: Dispatch<Group = GroupA>> _Kita0<GroupA, GroupA> for (T0, T1) {
-        const NAME: &'static str = "Blanket AA";
-    }
-    impl<T0: Dispatch<Group = GroupA>, T1: Dispatch<Group = GroupB>> _Kita0<GroupA, GroupB> for (T0, T1) {
-        const NAME: &'static str = "Blanket AB";
-    }
-    impl<T0: Dispatch<Group = GroupB>, T1, M1> _Kita0<GroupB, M1> for (T0, T1) {
-        const NAME: &'static str = "Blanket B*";
-    }
-    impl<T0: Dispatch<Group = GroupA>> _Kita1<GroupA> for T0 {
+    impl<_0: Dispatch<Group = GroupA>> _Kita0<GroupA> for _0 {
         const NAME: &'static str = "Blanket A";
     }
-    impl<T0: Dispatch<Group = GroupB>> _Kita1<GroupB> for T0 {
+    impl<_0: Dispatch<Group = GroupB>> _Kita0<GroupB> for _0 {
         const NAME: &'static str = "Blanket B";
     }
-
-    impl<T0, T1> Kita for (T0, T1) where T1: Dispatch, T0: Dispatch, Self: _Kita0<<T0 as Dispatch>::Group, <T1 as Dispatch>::Group> {
-        const NAME: &'static str = <Self as _Kita0<<T0 as Dispatch>::Group, <T1 as Dispatch>::Group>>::NAME;
+    impl<_0: Dispatch<Group = GroupA>, _1: Dispatch<Group = GroupA>> _Kita1<GroupA, GroupA> for (_0, _1) {
+        const NAME: &'static str = "Blanket AA";
     }
-    impl<T0> Kita for T0 where T0: Dispatch, Self: _Kita1<<T0 as Dispatch>::Group> {
-        const NAME: &'static str = <Self as _Kita1<<T0 as Dispatch>::Group>>::NAME;
+    impl<_0, _1> _Kita1<GroupB, GroupA> for (_0, _1) where _0: Dispatch<Group = GroupA>, _1: Dispatch<Group = GroupB> {
+        const NAME: &'static str = "Blanket AB";
+    }
+    impl<_0: Dispatch<Group = GroupB>, _1, _MŠČ0> _Kita1<_MŠČ0, GroupB> for (_0, _1) {
+        const NAME: &'static str = "Blanket B*";
+    }
+
+    impl<_0> Kita for _0 where _0: Dispatch, Self: _Kita0<<_0 as Dispatch>::Group> {
+        const NAME: &'static str = <Self as _Kita0<<_0 as Dispatch>::Group>>::NAME;
+    }
+    impl<_0, _1> Kita for (_0, _1) where _0: Dispatch, _1: Dispatch, Self: _Kita1<<_1 as Dispatch>::Group, <_0 as Dispatch>::Group> {
+        const NAME: &'static str = <Self as _Kita1<<_1 as Dispatch>::Group, <_0 as Dispatch>::Group>>::NAME;
     }
 };
 */